requirements.txt
absl-py==2.1.0
albumentations==1.3.1
annotated-types==0.5.0
anyio==3.7.1
appdirs==1.4.4
astunparse==1.6.3
bchlib==0.7
beautifulsoup4==4.12.3
bitarray==2.9.3
bitstring==4.1.4
BitVector==3.5.0
cachetools==5.5.0
certifi==2024.8.30
cffi==1.15.1
charset-normalizer==3.4.0
click==8.1.7
cryptography==44.0.2
cycler==0.11.0
deepface==0.0.93
exceptiongroup==1.2.2
fastapi==0.103.2
fastpbkdf2==0.2
filelock==3.12.2
fire==0.7.0
Flask==2.2.5
Flask-Cors==5.0.0
flatbuffers==24.3.25
fonttools==4.38.0
fuzzy-commitment @ git+https://github.com/cburkert/fuzzy-commitment.git@69749cf59093d374e78673afdeb8b906cad45f05
fuzzy-extractor==0.3
galois==0.4.4
gast==0.4.0
gdown==4.7.3
google-auth==2.36.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
gunicorn==23.0.0
h11==0.14.0
h5py==3.8.0
httptools==0.6.0
idna==3.10
imageio==2.31.2
importlib-metadata==6.7.0
itsdangerous==2.1.2
Jinja2==3.1.4
joblib==1.3.2
keras==2.11.0
kiwisolver==1.4.5
ldpc==0.1.45
libclang==18.1.1
llvmlite==0.39.1
Markdown==3.4.4
MarkupSafe==2.1.5
matplotlib==3.5.3
mtcnn==0.1.1
networkx==2.6.3
numba==0.56.4
numpy==1.21.6
oauthlib==3.2.2
opaque==1.0.0
opencv-python==4.6.0.66
opencv-python-headless==4.2.0.34
opt-einsum==3.3.0
packaging==24.0
pandas==1.3.5
Pillow==9.5.0
protobuf==3.19.6
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
pydantic==2.5.3
pydantic_core==2.14.6
pyldpc==0.7.9
pyparsing==3.1.4
PySide6==6.5.3
PySide6-Addons==6.5.3
PySide6-Essentials==6.5.3
PySocks==1.7.1
pysodium==0.7.18
python-dateutil==2.9.0.post0
python-dotenv==0.21.1
python-multipart==0.0.8
pytz==2024.2
PyWavelets==1.3.0
PyYAML==6.0.1
qimage2ndarray==1.10.0
qt6-applications==6.5.0.2.3
qt6-tools==6.5.0.1.3
qudida==0.0.4
reedmuller==1.1.2
requests==2.31.0
requests-oauthlib==2.0.0
retina-face==0.0.17
rsa==4.9
scikit-image==0.19.3
scikit-learn==1.0.2
scipy==1.7.3
seaborn==0.12.2
shiboken6==6.5.3
six==1.16.0
sniffio==1.3.1
soupsieve==2.4.1
starlette==0.27.0
tensorboard==2.11.2
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.11.0
tensorflow-estimator==2.11.0
tensorflow-io-gcs-filesystem==0.34.0
termcolor==2.3.0
threadpoolctl==3.1.0
tifffile==2021.11.2
tqdm==4.67.1
typing_extensions==4.7.1
urllib3==2.0.7
uvicorn==0.22.0
uvloop==0.18.0
watchfiles==0.20.0
websockets==11.0.3
Werkzeug==2.2.3
wrapt==1.16.0
zipp==3.15.0
